Add seed confirmation page (resolves MBS-11092)
With the switch to SameSite=Strict for login cookies, the experience of
seeding from external sites is degraded, because the user must login
again every time. (At least the POST parameters are preserved now, and
the username/password *might* be auto-filled, but it's still unclear to
the user why this is necessary.)

One way to make this /slightly/ nicer is to detect cases where the
Origin header differs from the MB origin, and direct the user to a
landing page where they can confirm the submission in a single click.
Since the new submission happens from MB's origin, any login cookies
stored by the browser will be sent with the "confirmed" submission.
(Unfortunately, the landing page will still show the user as logged out,
which I haven't thought of a good solution for but may not be a problem
in practice.) This also maintains the spirit of making the login cookies
first-party only, which was to block external scripts from submitting
forms without the user's knowledge or confirmation.

As an added feature, I've added to ability to see and modify what
parameters will be sent with the request.

I've tested the flow manually against my personal domain, where I have
an MB instance running behind SSL, and with a popular release importer
script modified to submit to my domain.