Remove copy-webpack-plugin
I'm having trouble with this plugin with Webpack v5 in watch mode. It's
currently only used to copy 3 files to the build dir:

 * root/favicon.ico
 * root/robots.txt.production
 * root/robots.txt.staging

So it can be replaced with a single cp in compile_resources.sh.

The only "regression" here would be that the files are not re-copied *in
watch mode when they change*, but (1) watch mode is for development,
where robots.txt serves no purpose, and (2) the favicon is one of the
least-likely resources to change.