MBS-11186: Add missing space between class names
When an editor has entered an edit and yet a vote type
(approval, since others are not allowed in this case)
applies to the same editor, votename doesn't separate
the applied classes, and adds the class "Approveowner"
which unsurprisingly does nothing.

Also removing extra spaces that were being added to the HTML
while at it.