Move CSRF tokens to form layer
CSRF tokens are now validated as part of the respective forms they're
expected by. This replaces the previous implementation which tied token
generation to the presence of the SecureForm attribute. This was
deficient because any action requiring a login can display the login
form if necessary, even if the original endpoint didn't use SecureForm.

There may have been a hacky way to have RequireAuth imply SecureForm,
but in general it makes more sense to tie the tokens to forms, since
forms can change endpoints or be used by multiple endpoints
(Form::User::Login, Form::Confirm, etc.), and there's no way to ensure
all of those endpoints use SecureForm as required.

Another thing that's been changed is the method of token generation.
Previously we used generate_token, which generates a random string using
system entropy. The new method is to reuse `generate_nonce`, which
doesn't require system entropy but is still very unlikely to be guessed.
I believe this is preferrable to wasting system entropy, because a new
token must be generated even when a form validates, as it's unclear if
the form will be used again. (For example, a user may enter login
credentials that pass form validation, but fail to authenticate; in this
case a new token is needed because the form will reload, but we can't
predict whether that will happen from inside the form.)