Allow PartialDate form files to work with state
These dispatches will eventually be sent up to the top form, but the
idea is that they will be sent back to these files so that the actual
actions don't need to be defined again and again.

We won't want to make the errors show immediately when the user types
the dates, because that'd be annoying (especially for the "less than 4
digits for the year" error, which would be hit by anyone typing the year
in by hand.  Other forms use debounce, but we can't use debounce inside
an async dispatch call. As such, what this does is adding a
"pendingErrors" array that stores the errors calculated on the fly and
only actually moving them to the displayed 'errors' field when we decide
we want to show them (usually on field blur).