Convert date_range_fieldset to React
Partially based on work by Anirudh Jain.

This converts date_range_fieldset to React using state.

These dispatches will eventually be sent up to the top form, but the
idea is that they will be sent back to these files so that the actual
actions don't need to be defined again and again.

This adds one new feature not present in the TT version: the ability to
copy the starting date to the end date with one click, as already
present in the relationship editor.