MBS-11168: Convert add/edit alias form to React
Partially based on work by Anirudh Jain.

We don't allow submitting this if any errors are pending. This also
blocks submitting if there are *pending* errors, to avoid someone
submitting before the errors were shown.

Disabling the submit button also disables submission on pressing Enter,
but just in case, we ensure the form can't be submitted in onSubmit, and
we ensure we display the errors on trying to submit pressing Enter if
said errors exist.