MBS-9388: List relationships credits in alias tab
We display all other names given to the entity in the DB here,
it probably makes sense to display these too.

Added a new query that doesn't join anything rather than reusing
the existing relationship loading query, since that should make this
a lot faster.

Eventually it would make sense to make these credits link to a page
with all the relationships they are used in, like we do with artist
credits now. But this might be a start already to help supplement
aliases.