Better UI for disabling fields for search hints
Many fields are irrelevant for search hints, so we have always disabled
them for submission when the "search hint" type is selected. This can be
confusing, given that the values are still displayed in the disabled
fields - additionally, if any of the disabled fields had errors, the
errors would still block submission even if no longer relevant for the
form.

This blanks the fields when the user selects the search hint type, and
saves them in state to bring them back if the user selects a different
type again.