Add csrf_session_key field
Currently, the session key under which we store a form's CSRF token is
just a JSON string containing details about the form and request. This
means that if you load multiple instances of the same form in separate
tabs, only the form that was loaded last will allow submission on first
attempt. But it was necessary that we were able to generate the same
session key when the form was first loaded, and when the form was
submitted (in order to retrieve the expected token from the session).

The implementation here improves the scenario when opening multiple tabs
by adding an additional hidden input named csrf_session_key. Like
csrf_token, this is a nonce generated in the same way, but indicates
where the token is stored in the session object. The new keys expire
after 10 minutes. Since each token is stored in a unique key per
request, separate tabs' tokens will no longer clobber each other.