MBS-11216: Split InstrumentsWithoutAnImage report
Wikidata is not connected to images at all anymore,
but it's still a very useful link to have. As such, it makes sense
to split this report in two.
Also changed the query a bit to hopefully improve it.

We could just remove InstrumentsWithoutAnImage, but hopefully
we can soon use it again once we start being able to link
to IROM images on StaticBrainz.