Refactor ReportsIndex to reuse component for each entry
This slightly shortens the file, but more importantly,
it enormously limits the amount of changes we'll need to make
if we decide to change how we display these and it limits
the chances of errors linking to the report (which I've seen
happen in PRs) by just asking for name and content.

I could see no reason why the report from MBS-10843
used display: none rather than just being commented out,
so I commented it out.