MBS-11217: Only show instrument reports to relationship editors
Only relationship editors can do anything about these,
so there's no reason to link to them from ReportsIndex
unless the user is a relationship editor.

We can leave the actual content of the report public,
unlike with LimitedEditors, in case a relationship editor
wants to specifically share a link to it (say, to show which
instruments could use a new image).