MBS-10634: Limit InstrumentsWithoutAnImage to StaticBrainz
Commons images (and other image links) are pretty much
useless to us at the moment, since we no longer display any of them.
This limits the InstrumentsWithoutAnImage report to only count
staticbrainz.org/irombook images, since the point is to see
which instruments are missing an image we can display.