MBS-11229: Convert Edit Cover Art edit to React
There's no such thing as position in the display data at all
on the Perl side, so removing that from the template too
since it will never be there.
Also, we might as well pass the types and get the translated names
on the JS side rather than just using l_name on Perl,
which we don't do elsewhere either (admittedly, elsewhere
we pass the English titles only, but I feel this is better).