Fix performance issues with MBS-1424 implementation
After running the original upgrade script for MBS-1424 in production we
started to observe an increase in 504s due to track insertions taking
2-4 seconds or more. I traced this to the recording_release_dates view
being slow, particularly joining against a union of the release_country
and release_unknown_country tables.

This commit adds a new materialized `release_first_release_date` table
to store a union of the earliest release dates from `release_country`
and `release_unknown_country`.

To see where this matters, refer to the changes in the
`set_recordings_first_release_dates` function. An example of the
previous kind of subquery this would perform is:

SELECT DISTINCT ON (recording)
       recording, year, month, day
  FROM recording_release_dates
 WHERE recording = ANY(ARRAY[1076099])
 ORDER BY recording, year NULLS LAST, month NULLS LAST, day NULLS LAST;

which takes slightly under 4 seconds in production. The new subquery
(using the release_date table) looks like this:

SELECT DISTINCT ON (track.recording)
    track.recording,
    rd.year,
    rd.month,
    rd.day
  FROM track
  JOIN medium ON medium.id = track.medium
  LEFT JOIN release_first_release_date rd ON rd.release = medium.release
 WHERE track.recording = ANY(ARRAY[1076099])
 ORDER BY
  track.recording,
  rd.year NULLS LAST,
  rd.month NULLS LAST,
  rd.day NULLS LAST;

and takes only ~0.66 ms by making use of an index scan on
release_first_release_date_pkey.