Use SameSite=Lax for login cookies (fixes MBS-11238)
5ab8cfba261175f4712d89a230b2c8b0c2bc716a made it so that session/login
cookies are only sent in first-party contexts from the same origin. This
means the user will appear as logged out if they click on any links to
MB from an external origin. In order to address this, we either have to
downgrade to SameSite=Lax and ensure we don't allow any data
modifications via GET requests, or add a separate non-strict cookie that
points to a user session for display but doesn't authenticate as that
user. I'm not sure how to do the latter with Catalyst without breaking a
lot of assumptions we make throughout the codebase, so I'm "downgrading"
the cookies to Lax mode and tweaking a couple cases where it might be
possible to submit changes via a GET request. This is still an
improvement over the pre-5ab8cfb behavior of SameSite=None and fulfills
the spirit of that commit.