Rename EditorT to UnsanitizedEditorT
This should discourage use of EditorT in favor of SanitizedEditorT.
While this doesn't affect what data is returned from the server, I think
it serves as a good reminder to check whether a full editor is really
needed. In a followup commit I plan to make Editor::TO_JSON return a
SanitizedEditorT by default.

This commit also removes the CatalystUserT alias for EditorT to maintain
consistency in indicating that the type is unsanitized.