MBS-11250: Properly support JSON user-rating lookups
Asking for this as JSON didn't error, but serialized the entity instead.
This changes it so that you can actually query for the rating as in XML.

Note this returns the rating in a 0-100 scale, while normally user-rating
is returned on the 0-5 scale. This is inconsistent with the rest of MB
but consistent with the XML version of this endpoint, so probably
for the best.