MBS-8700: List genres in JSON-LD
This uses only the genres on top_tags (so, the ones that would be
displayed on the sidebar). This sounds sensible enough to me,
since if someone has added 100 genres to an artist we probably don't
want to show them all.

https://schema.org/genre specifically allows only Text or URL for genre,
so I went with URL to at least be able to pass an MBID for it. Ideally
we would serialize a genre object with ID and name, but that doesn't
seem to be supported.