Set SameSite=None on beta cookie
If someone has the beta preference on and uses an external importer, the
POST request won't include the beta cookie due to it using SameSite=Lax.
This switches the beta cookie to use SameSite=None so that the 307
redirect is performed before the ConfirmSeed page is shown twice (first
on the production site, then again on the beta site).