MBS-9674: Support internationalized domains in URL forms
This was failing because, unsurprisingly, it couldn't find the
ASCII hostname in the Unicode URL. I used punycode (which we
already installed as a dependency, but I added as a requirement)
to decode the punycode back into the original unicode URL
for hostname comparison.

During a Monday meeting we also agreed to always display
international URLs as Unicode (which we already did when prettifying
but not in the links editor). This makes the relevant changes
to externalLinks and to URL forms.