MBS-11256: Make adding tracklist to empty medium an autoedit
We generally make adding something where nothing existed an autoedit.
I see no reason why this shouldn't also apply to adding a tracklist if it was
previously unknown.

The first medium edit test already added a new track to an empty tracklist,
so testing autoeditness with that and non-autoeditness of further adds
with the second track addition further down.