MBS-4548: Allow seeing all uses of an artist credit
Currently to see where an artist credit is used you need to visit
three different tabs of an artist page (overview, releases and
recordings) and filter each one by AC. And you still aren't able
to see ACs that are used on tracks only.

This mostly apes the tags interface to show a list of artist credit
uses. The main use for this is to either check it before entering
an edit artist credit edit (from the aliases page) or while voting
on one, so I added links from those two cases. For the edit AC edit
the link is only shown while the edit is open for voting, since
after that it's of no use.