MBS-11267: Always show artwork info when adding/reordering
While generally you can see the types and comment for a specific
image when hovering over the image, that's not too clear in general and
it doesn't help at all when the thumbnail hasn't been generated yet
(since there's no image to hover over).

This adds an information icon that displays the data on hover, and also shows
if there's no thumbnail yet.

Additionally, we weren't loading types for add_cover_art.