MBS-10208 (II): Allow merging collections (Implementation)
This allows merging collections of the same entity type,
moving all entities to the destination collection.
All collaborators for the collections being merged will be collaborators
in the destination collection. The descriptions of all collections
will be appended (as with annotations in entity merges) for the user
to later edit them as desired.

For each entity that appears in more than one collection, the oldest
value is kept for "added", and values for "comment" are concatenated
like for annotation merges. The value for position is kept as the
destination value (or, if all dupes are on non-destination collections,
it is chosen at random from one of the collections). This can be
changed if user feedback requires it once the position field start
actually being used.

To stop a user from being able to see info about other users'
private collections by entering random IDs into the merge queue URL,
the merge queue detaches to 403 if passed a collection with
a different owner.