MBS-11181: Don't merge relationships with different ended data
If you have a non-ended relationship with a start date and you try
to add a second relationship without dates but marked as ended
the code currently marks the existing relationship as ended rather than
adding a second relationship.
This might have been intentional, but I'd argue it's confusing (the user
who tried to add the second relationship might not realize
what happened) and makes an assumption that might not always
be correct. In the example given for the ticket, the relationship
- membership - ended at some unknown time, then started again
from the given date and still applies at present.
While making this change might be a bit more annoying for cases where
the user did intend the change to apply to the existing relationship,
doing it like that means you can pick whether to edit the existing
relationship or add a new one, while the current way just blocks
one of those options unless you add a spurious date just to have
the rel be allowed, then remove it to leave just "ended".