MBS-11183: Always preserve merged entity name as alias
The code always added the name of any entities being merged
as an alias to the destination entity, with one exception:
if the same name was already used as an alias for any other entity
of the same type.
This probably dates back to when we only allowed each alias
to be used once across the DB, but we stopped doing that ages ago.
As such, the commit changes it to only skip adding the name as an alias
if the destination entity *already* has that name as an alias.