MBS-11289: Stop autocleaning YouTube Music -> YouTube
YouTube Music links shouldn't be converted to YouTube links,
since they're intended to be a different thing and, in some cases,
are not available to anyone else to YouTube Music subscribers.

For now I'm not making them autoselect paid or free streaming, since
most of them *can* be streamed for free with ads, but some
are strictly limited to paid streaming only.