MBS-11292: Do not use UN flag for [Worldwide]
As suggested in the ticket, that's not quite appropriate.
This uses a globe flag - sadly, it gets so distorted at this tiny size
that it is hard to understand what's going on, but that was also
completely true of the previous one.

The flag was made by user pmepepnoute trying to keep a similar color
to the previous one.