Use artist name for empty AC name in /ws/js/edit
Prior to this commit, if the release editor somehow failed to populate
an artist credit name field (due to a bug like MBS-10964), submitting it
as an empty string, the resulting edit created would ultimately fail due
to violating a non_empty DB constraint on the name field.

This works around the issue by falling back to the artist name if the
credited name is empty. I've added this behavior to both the client and
server.

You'll notice that some incidental improvements have been made to
`process_artist_credits` related to id/gid loading. Previously, we only
checked for and loaded gids to fill in numeric ids where they were
missing. Since we now want to fill in the name if empty for *all*
artists, we load all artists, regardless of the type of id, and do that.
This is in addition to filling in the numeric id as before -- and the
new code supports an additional case where an artist submitted with only
a numeric id will have a gid added too.