MBS-4782 / MBS-11333: Pass original MBID to AddCoverArt if no release
We store the original MBID, so we should use it. The TT used to use
the original MBID straight from the data (not passing it
to display data) for the filename display, but didn't use it for the
actual cover art display, so that was broken. Passing the MBID to the
Release->new call ensures both filename *and* image work.