MBS-11059: Indicate which release groups have CAA art in listings
This adds a new 'load_has_cover_art' method and a new 'has_cover_art'
boolean to be able to have a quicker query for this that doesn't
actually load all the artwork for the release group.

Since 'has_cover_art' was already the predicate for 'cover_art',
I changed that to 'has_loaded_cover_art'. I also renamed that
for releases, since it seems more consistent that way.