MBS-11337: Always link to URL entity page and link on URL reports
On reports about URLs, it makes sense to always link to the URL entity
(since that's where you can fix the problems). It still makes sense
to have a link to the actual destination though.
At the moment UrlRelationshipList showed URLs in the same way
as elsewhere on the site (destination link on click + [info] link
to the entity) while the other lists being changed here only linked
to the entity but did not allow following the link.
This makes both of these show both the URL link and a link to the
URL entity (shown as the MBID), which allows visiting both links
as needed without having to click the tiny info link that moves
line by line depending on URL length.