MBS-11344: Validate rateyourmusic URLs
These are all the cases that clearly map to ours. There's a few
other types of URLs that have sometimes been linked to our
entities:
User pages (/~) have been linked to artists, but probably should not be.
/film and /films pages are about films, but sometimes
seem to get linked to music artists and whatnot. I'm not currently
accepting them, we can see if people think we should allow them
and for what entity types.
/list pages are user-created lists, they have been linked to different
places but that doesn't seem right.
/wiki pages are, well, that. It seems like some people use them
for distributors. I'm currently blocking them but they could be
allowed with some conditions in the future if the community asks for it.