MBS-6502: Surround catno search with quotes
CRD 3443 is a valid catalog number we should support,
and in fact our search supports it. But by sending 'catno:CRD 3443'
we actually search 3443 as a title!

Since there's one use case that quotes break (searching for a regex,
where the string must start and end with slashes), we avoid quoting
any query that starts and ends with slashes. We also avoid double
quoting any query already starting and ending in quotes.