MBS-11360: Work around unconsistent name for RG type
For some reason, the entity_name for release group is 'rg'
in Controller::ReleaseGroup, despite its ENTITIES entry being
'release_group'. This seems very weird to me but it'd probably be a mess
to change it now, so I just massage it into what the JS code expects
instead so that the RG alias form will work.