MBS-11366: Also grey out harmful links after URL entity is deleted
We still show the links in, say, edit history, so we should make sure
they are still disabled. I see no reason why we would ever *not* want
to disable a deleted link that is in the harmful list, so I didn't pass
the disableLink boolean down and just hardcoded it to always happen.