MBS-11369: Revert to middle vertical alignment for table.details data (#1898)
Detail table rows begin with colon-terminated headers which can easily
wrap onto multiple lines. When this occurred, the value associated with
the header (which follows in the next cell) could appear far above
where the header ends when the value's cell was top-aligned,
potentially making the value appear unrelated to the header and harder
to parse than it needed to be.