MBS-11354: Don't show arrows unless we want ordering (#1885)
This was changed with 72ce47118a42097872d4a15bb2407e1c9d14c9b9
in a way that caused the arrows to always show for orderable
relationship types. It used to only show once there were at least
two relationships and the user selected that they wanted them
to be ordered (hasOrdering). The old way seems more sensible to me.