MBS-11375: Use top-align for release event tables in Edit release
MBS-11369 changed the vertical align and in most places
that's for the best, but release event diff tables are better
with top align. Otherwise, a long list being turned into a small one
(e.g. edit #74086408) has the smaller value way down the page and makes
understanding the change a lot harder.