Remove unaccented_name, perform unaccent on client
These fields take up a lot of space in our entity JSON. I believe we can
significantly reduce our markup size (from hydrated element props) in
some cases by removing them.

I didn't check that the JavaScript version is equivalent (it's most
likely not), but I don't think it has to be perfect for what we're using
it for.