Add /ws/js/entities
This endpoint allows looking up multiple MBIDs or IDs in a single
request. My current use case for this is the "recent entity" lists in
the autocomplete. Currently those are stored as complete JSON objects in
localStorage, but that can produce unrecoverable errors if the expected
JSON format changes from what is stored (MBS-10498). (Or the JSON may be
valid, but just out of date.) Instead, we'll store the MBIDs in
localStorage and fetch the actual entity JSON in the background after
the page loads.