MBS-11349: Make medium length consistent
We currently take medium length from the track durations, if loaded.
If tracks are not loaded, we check if we can get the duration
from the discIDs.
DiscID duration ignores pregap and data tracks though, so the two
values are sometimes different.
This uses the new medium_track_durations view to always provide
consistent medium durations.

The test in Entity::Medium is not useful for the new calculation
method, and there's a test in Data::Medium that already checks
whether length calculation works (and which no longer needs
to load the tracks to work).