MBS-10343: Specify whether cover art has been darkened
We should specify why we don't have cover art for darkened
releases, rather than just say "No cover art available". Otherwise,
this can make users go to the cover art tab to add cover art,
only to be surprised to find out that it is not allowed.
Darkened status isn't currently being updated properly,
but at least with this it will show up as it should once that is fixed.