MBS-11394: Don't try to show cover art for darkened releases
For darkened releases, cover art is not available, so trying to show it
will only cause one or several broken image boxes.
This replaces the content of the cover art tab with a notice that the art
has been taken down. As such, there's also no need to show a specific
"can't upload art" section anymore.

Note: for this to be truly useful we will need to implement MBS-6567.