MBS-10470: Rename "IRC" link in footer to "Chat (IRC)"
Suggestion was to make this "Chat" because it is more clear
to new users. While that is probably true, other users might still
search for "IRC", so I feel "Chat (IRC)" is short enough, clear enough
for newbies and still useful for people who are searching for an
IRC link.