MBS-9987: JSON-LD: Use "Concept URI" for Wikidata
These are the IDs Wikidata itself uses for RDF / SPARQL,
so it makes linked data usage a lot easier by avoiding the user
having to put the right URI together.
They redirect to the standard page when loaded in the browser,
so it shouldn't make things harder for data users otherwise.