MBS-9884: Use "Created" for character-type artist begin dates
The documentation says "Begin date represents the date (in real life)
when the character concept was created." As such, "Created" is the best
option for this, much better than "Begin date".
The docs also say "The End date should not be set", so no point
setting a specific label for that.