MBS-9876: Stop redirecting tag search to direct search
We support searching for tags with SOLR, so we should probably
allow doing that. Because the response does not include tag ID,
we just load the tag based on the returned name.

For what it's worth, the results are currently significantly worse
than with direct search. We might want to try and improve that.