MBS-9857: Make merging HTTP->HTTPS versions of URL an auto edit
We generally want to make it easy to turn HTTP URLs into HTTPS ones.
We already had code to allow auto-editing an URL if the change
was only HTTP->HTTPS and it did not cause a merge.
But in fact, there's no reason not to allow this also for a merge,
since the only reason such a change is incorrect is if the HTTPS URL
is wrong, in which case the URL can easily be changed back
to HTTP post-merge (there's no case I can imagine where we want
to have both HTTP and HTTPS versions and keep some relationships
using one version and some using the other).