MBS-9826: Show more realistic edit closing time overviews
It's quite misleading to say "closes in 6 days" as soon as an edit
is entered, given that it's really 6 days, 23 hours and 50-something
minutes. Yes, we have a tooltip with the date, but still.
Instead of using floor, this uses round - it seems more sensible
to say 7 days when it's more than 6 and a half, etc.
I start displaying hours once they hit 23 (rounded, so at 23 h and
30 minutes left), rather than still saying 1 day until there's only
12 hours left like elsewhere.