Fix trimming annotation field for edited release
The release editing form’s preview was unintentionally not processing
the annotation field. That was hidden by the patch for MBS-11317 that
fixed the issue client-side only, so not solving the issue in /ws/js.

This patch makes /ws/js to use the same trimming code as annotation
editor, so that it removes trailing spaces from annotation in release
editor the same as from annotation in annotation editor.

Add test for annotation trimming in /ws/js (used by release editor).
Test annotation text is similar to text in AddAnnotation tests.