MBS-11427: Improve seed-love-bug test/showcase
This is not particularly good as a showcase of fields to seed.
Not only is it missing several available fields (and not showing
usage of most of the possible multi-option combos)
but it has no useful labeling for fields and no placeholders,
making it very hard to even understand what's going on
(several MBID fields are listed one after another with no info
on what MBIDs are being seeded!).

This makes it so that every field or related set of fields has a proper
legend, and the most confusing ones also have a placeholder specifying
what should (or can) be entered there.