MBS-10367: Allow requesting copy of sent report when reporting
This allows a reporter to save a copy of the reports they have
sent, if so desired. That might be helpful to, for example,
remember if you have already reported a user recently but the
report still hasn't been acted on, and as such avoid sending
another report on the same issue.

In order to reuse the actual content of the report email
in the copy, this moves the message by the editor, if any,
to be above the "you can reply to this" or whatnot signature.
That seems sensible anyway - all the content should precede that.