MBS-9836: Guess Case: stop uppercasing "the" in artist names (#1951)
There's no documentation about why this was done this way
in the first place, but most likely it was because of pre-NGS artists
which would have two or more different, full artist names as part
of a collaboration artist name. This is fairly uncommon with NGS,
while artists which just have "the" in their names
(think "Tyler, the Creator") and where "the" should be lowercase
are reasonably common.
I asked the community and most feedback suggested dropping
this unexplained / unexpected difference between artist guess case
and guess case for other entities.